requirements.txt
langchain
langgraph
langchain-core
langchain-chroma
langchain-community
langchain-openai
openai
chromadb
pandas
streamlit
tiktoken
python-dotenv
ipykernel
serpapi
langchain-graph
google-search-results